isolate examples to another file
